+++ tests/basic_tests/app/tests.py
@@ -28,8 +28,21 @@ class BasicTest(TestCase):
         pass
 
     def test_insalled_apps(self):
+        installed_apps = list(settings.INSTALLED_APPS)
+        # remove unnecessary library
+        unnecessary_library_list = (
+            'django_jinja',
+            'debug_toolbar',
+            'coffin',
+        )
+        for x in unnecessary_library_list:
+            try:
+                installed_apps.remove(x)
+            except ValueError:
+                pass
+
         self.assertEqual(
-            settings.INSTALLED_APPS, [
+            installed_apps, [
                 'app', 'app2', 'django.contrib.auth',
                 'django.contrib.contenttypes', 'django.contrib.messages',
                  'django.contrib.sessions', 'django.contrib.admin',
@@ -53,9 +66,25 @@ class BasicTest(TestCase):
         self.assertEqual(response.context["the_answer"], 42)
         self.assertTemplateUsed(response, "test1.html")
         # on windows, newline is CRLF.
-        self.assertIn(response.content, [b"<h1>test1: 42</h1>\n", b"<h1>test1: 42</h1>\r\n"])
+        self.assertEqual(response.content.replace('\r', ''), b"<h1>test1: 42</h1>\n")
         response = c.get("/testnotfound/")
         self.assertEqual(response.status_code, 404)
+
+    def test_view_with_jinja(self):
+        from importd import COFFIN, DJANGO_JINJA
+        if not COFFIN and not DJANGO_JINJA:
+            # cannot not jinja2 test if jinja2 integration is not exist
+            return
+
+        c = Client()
+        response = c.get("/test2/")
+        self.assertEqual(response.status_code, 200)
+        if COFFIN:
+            # with django-jinja, cannout use response.context, assertTemplateUsed
+            # but with coffin, it works.
+            self.assertEqual(response.context["sample_list"], range(3))
+            self.assertTemplateUsed(response, "test2.jinja")
+        self.assertEqual(response.content, b"<h1>test2: 0|1|2</h1>")
 
     def test_mounts(self):
         c = Client()